front/syn_metadata: move collect_program_metadata + collect_fields_and_returns + collect_nested_fn_returns out of front/ast (Slice 2)
