+++ majit/majit-translate/src/front/ast.rs
@@ -4,13 +4,10 @@
 //! This module lowers syn AST nodes into FunctionGraph ops with proper data flow (Variable linking).
 
 use std::collections::HashMap;
-use syn::Item;
 
-use crate::ParsedInterpreter;
-use crate::flowspace::model::ConstValue;
 use crate::model::{
-    BlockId, CallTarget, ExitSwitch, FrameState, FunctionGraph, ImmutableRank, Link, LinkArg,
-    OpKind, UnknownKind, UnsupportedExprKind, UnsupportedLiteralKind, ValueType, exception_exitcase,
+    BlockId, CallTarget, ExitSwitch, FrameState, FunctionGraph, Link, LinkArg, OpKind, UnknownKind,
+    UnsupportedExprKind, UnsupportedLiteralKind, ValueType, exception_exitcase,
 };
 
 pub use crate::front::semantic::{
@@ -27,7 +24,7 @@ use crate::front::syn_metadata::{
     binary_result_value_type_inner, canonical_pat_name, cast_builtin_name, const_value_value_type,
     dyn_trait_root_from_type_str, extract_dyn_trait_root_with_context,
     extract_element_type_from_str, intrinsic_call_result_type, is_synthetic_ctor_path,
-    is_synthetic_result_option_wrapper_path, kind_char_to_value_type, matches_constructor_path,
+    matches_constructor_path,
     method_as_ref_return_type, op_result_value_type, outer_generic_inner_type,
     path_as_value_float_constant, split_tuple_type_elements, transparent_option_inner_type,
     transparent_result_err_type, type_root_from_type_string, type_root_ident,
@@ -124,454 +121,7 @@ macro_rules! get_value_var {
 /// another file — single-file metadata leaves cross-file calls
 /// (`!crate::is_str(...)` from dictobject.rs against `is_str`
 /// defined in strobject.rs) unclassified.
-pub fn collect_program_metadata_pub(parsed_files: &[ParsedInterpreter]) -> ProgramMetadata {
-    let mut known_struct_names = std::collections::HashSet::new();
-    let mut known_trait_names = std::collections::HashSet::new();
-    let mut struct_fields = StructFieldRegistry::default();
-    let mut fn_return_types: HashMap<String, String> = HashMap::new();
-    let mut immutable_fields: HashMap<String, Vec<(String, ImmutableRank)>> = HashMap::new();
-    let mut struct_origins: HashMap<String, String> = HashMap::new();
-    let mut use_imports: HashMap<(String, String), String> = HashMap::new();
-    let mut module_statics: HashMap<(String, String), crate::parse::ModuleStaticDecl> =
-        HashMap::new();
-    for parsed in parsed_files {
-        crate::front::syn_metadata::collect_struct_names(
-            &parsed.file.items,
-            "",
-            &mut known_struct_names,
-        );
-        crate::front::syn_metadata::collect_trait_names(
-            &parsed.file.items,
-            "",
-            &mut known_trait_names,
-        );
-        // PyPy `annrpython.py` bookkeeper: every newly-seen STRUCT
-        // gets cached under its lltype-object identity, which is the
-        // defining-module path.  Pyre carries names as strings — record
-        // `bare_name → module_path` so cross-file references can
-        // resolve to the canonical hash slot the runtime publishes.
-        // Empty `module_path` (legacy `parse_source` entry) skips the
-        // record; consumers fall back to dual-publish convergence.
-        if !parsed.module_path.is_empty() {
-            crate::front::syn_metadata::collect_struct_origins(
-                &parsed.file.items,
-                &parsed.module_path,
-                &mut struct_origins,
-            );
-        }
-        // Mirror the per-file `ParsedInterpreter.use_imports` into the
-        // program-wide `(module_path, alias) → fully_qualified_path`
-        // registry.  Caller may pass the same alias across multiple
-        // files; the per-file key disambiguates.
-        for (alias, full) in &parsed.use_imports {
-            use_imports.insert((parsed.module_path.clone(), alias.clone()), full.clone());
-        }
-        for ((nested, name), decl) in &parsed.module_statics {
-            let module = qualify_module_path(&parsed.module_path, nested);
-            module_statics.insert((module, name.clone()), decl.clone());
-        }
-    }
-    for parsed in parsed_files {
-        collect_fields_and_returns(
-            &parsed.file.items,
-            "",
-            &parsed.use_imports,
-            &known_struct_names,
-            &known_trait_names,
-            &mut struct_fields,
-            &mut fn_return_types,
-            &mut immutable_fields,
-        );
-    }
-    ProgramMetadata {
-        known_struct_names,
-        known_trait_names,
-        struct_fields,
-        fn_return_types,
-        struct_origins,
-        use_imports,
-        module_statics,
-    }
-}
-
-/// Walk all top-level (and nested `mod`) `Item::Struct` declarations in
-/// `items` and record each struct's bare name → defining module path.
-/// Mirrors PyPy `bookkeeper.getdesc(TYPE)` resolution: every observed
-/// lltype STRUCT identity has a canonical home module; pyre carries
-/// names as strings so this map serves the same role.
-///
-/// Nested `mod foo { struct Bar; }` extends the prefix to `outer::foo`
-/// so the registered origin matches what `path_hash(canonical)` would
-/// produce for the qualified key.
-/// Pass 1b: collect field types + fn return types using known_struct_names.
-///
-/// `use_imports` is the per-source `use <path> as alias` table — same
-/// map that lowering reads through `GraphBuildContext.use_imports`, so
-/// struct field type / fn return type metadata strings produced here
-/// land in the same name namespace as the parameter / local-binding
-/// type strings later mint by `qualify_type_name_with_imports`.  Without
-/// this thread-through, `bookkeeper.getdesc`-style alias resolution
-/// would diverge between metadata + lowering (PyPy single-frame
-/// `f_globals` walk: `rpython/annotator/bookkeeper.py:353`).
-fn collect_fields_and_returns(
-    items: &[Item],
-    prefix: &str,
-    use_imports: &HashMap<String, String>,
-    known_struct_names: &std::collections::HashSet<String>,
-    known_trait_names: &std::collections::HashSet<String>,
-    struct_fields: &mut StructFieldRegistry,
-    fn_return_types: &mut HashMap<String, String>,
-    immutable_fields: &mut HashMap<String, Vec<(String, ImmutableRank)>>,
-) {
-    for item in items {
-        match item {
-            Item::Struct(s) => {
-                let bare_name = s.ident.to_string();
-                // RPython: T.TO gives the actual lltype object.
-                // qualified_full_type_string uses known_struct_names to identify
-                // which inner types are user structs (not heuristic).
-                let fields: Vec<(String, String)> = s
-                    .fields
-                    .iter()
-                    .filter_map(|f| {
-                        let field_name = f.ident.as_ref()?.to_string();
-                        let field_type = qualified_full_type_string_with_imports(
-                            &f.ty,
-                            prefix,
-                            use_imports,
-                            known_struct_names,
-                            known_trait_names,
-                        )?;
-                        Some((field_name, field_type))
-                    })
-                    .collect();
-                // RPython: `_immutable_fields_ = ['a', 'b']` on the class
-                // body. We accept `#[jit_immutable_fields(a, b)]` on the
-                // struct declaration (proc-macro pass-through in
-                // `majit_macros::jit_immutable_fields`). Multiple
-                // attributes accumulate.
-                let immutables = crate::front::syn_metadata::collect_immutable_field_attrs(&s.attrs);
-                if !immutables.is_empty() {
-                    if prefix.is_empty() {
-                        immutable_fields
-                            .entry(bare_name.clone())
-                            .or_default()
-                            .extend(immutables.iter().cloned());
-                        // Canonical defining-module alias: `field_immutability`
-                        // callers route owner through `qualify_type_name`
-                        // non-empty-prefix canonical path; mirror the
-                        // `struct_fields` dual-publish below so the lookup
-                        // hits exactly under both spellings.
-                        let canonical = majit_ir::descr::canonical_struct_name(&bare_name);
-                        if canonical != bare_name {
-                            immutable_fields
-                                .entry(canonical)
-                                .or_default()
-                                .extend(immutables.iter().cloned());
-                        }
-                    } else {
-                        let qualified = format!("{}::{}", prefix, bare_name);
-                        immutable_fields
-                            .entry(qualified)
-                            .or_default()
-                            .extend(immutables.iter().cloned());
-                        immutable_fields
-                            .entry(bare_name.clone())
-                            .or_default()
-                            .extend(immutables.iter().cloned());
-                    }
-                }
-                if prefix.is_empty() {
-                    // Dual-publish under bare name + canonical defining-module
-                    // form when `STRUCT_ORIGIN_REGISTRY` (populated by
-                    // `analyze_pipeline_from_parsed:372`) supplies an
-                    // origin.  Use-site lookups now route through
-                    // `qualify_type_name` non-empty-prefix canonical path
-                    // (PyPy `bookkeeper.getdesc` analog) and land on the
-                    // canonical key directly without falling back to the
-                    // `unique_suffix_owner_key` shim.
-                    let canonical = majit_ir::descr::canonical_struct_name(&bare_name);
-                    struct_fields
-                        .fields
-                        .insert(bare_name.clone(), fields.clone());
-                    if canonical != bare_name {
-                        struct_fields.fields.insert(canonical, fields);
-                    }
-                } else {
-                    let qualified = format!("{}::{}", prefix, bare_name);
-                    struct_fields.fields.insert(qualified, fields);
-                }
-            }
-            Item::Fn(func) => {
-                // RPython: op.result.concretetype — module-qualified return type.
-                let ret_ty = match &func.sig.output {
-                    syn::ReturnType::Type(_, ty) => qualified_full_type_string_with_imports(
-                        ty,
-                        prefix,
-                        use_imports,
-                        known_struct_names,
-                        known_trait_names,
-                    ),
-                    syn::ReturnType::Default => Some("()".to_string()),
-                };
-                if let Some(ret_ty) = ret_ty {
-                    let key = if prefix.is_empty() {
-                        func.sig.ident.to_string()
-                    } else {
-                        format!("{}::{}", prefix, func.sig.ident)
-                    };
-                    fn_return_types.insert(key, ret_ty);
-                }
-                // Nested `fn`s declared as `Stmt::Item(Item::Fn(_))`
-                // inside this fn's body.  Rust's lexical scoping makes
-                // them callable only from within the parent body, but
-                // the classifier (`expr_unary_not_operand_kind`) sees
-                // them as bare-name `Expr::Call` paths and needs the
-                // signature to disambiguate `!nested_pred(arg)` between
-                // UNARY_NOT and UNARY_INVERT.  RPython parity:
-                // `bookkeeper.getdesc(value)` resolves any callable in
-                // scope by host-identity (`annrpython.py` callee
-                // resolution); pyre's static walker substitutes by
-                // registering the nested signature under the bare ident.
-                collect_nested_fn_returns(
-                    &func.block.stmts,
-                    prefix,
-                    use_imports,
-                    known_struct_names,
-                    known_trait_names,
-                    fn_return_types,
-                );
-            }
-            // RPython has no `Item::Const` analogue — Python module-level
-            // constants reach `flowcontext.py` as `LOAD_GLOBAL(name)`
-            // followed by the bookkeeper's PBC table lookup
-            // (`bookkeeper.py:329-340 immutablevalue`).  Pyre's walker
-            // doesn't model PBCs, so consts surface as plain
-            // `Expr::Path` identifier references; the classifier needs
-            // a typed-name entry to resolve `!FOO`-shape uses.  Reuse
-            // `fn_return_types` keyed by ident — Rust convention
-            // (SCREAMING_SNAKE_CASE consts vs snake_case fns) keeps the
-            // namespaces separate in pyre's source.
-            Item::Const(c) => {
-                if let Some(ty) = qualified_full_type_string_with_imports(
-                    &c.ty,
-                    prefix,
-                    use_imports,
-                    known_struct_names,
-                    known_trait_names,
-                ) {
-                    let key = if prefix.is_empty() {
-                        c.ident.to_string()
-                    } else {
-                        format!("{}::{}", prefix, c.ident)
-                    };
-                    fn_return_types.insert(key, ty);
-                }
-            }
-            Item::Impl(impl_block) => {
-                let self_ty_root = type_root_ident(&impl_block.self_ty);
-                for sub in &impl_block.items {
-                    if let syn::ImplItem::Fn(method) = sub {
-                        let ret_ty = match &method.sig.output {
-                            syn::ReturnType::Type(_, ty) => {
-                                qualified_full_type_string_with_imports(
-                                    ty,
-                                    prefix,
-                                    use_imports,
-                                    known_struct_names,
-                                    known_trait_names,
-                                )
-                            }
-                            syn::ReturnType::Default => Some("()".to_string()),
-                        };
-                        if let Some(ret_ty) = ret_ty {
-                            if let Some(ref ty_root) = self_ty_root {
-                                let qualified_ty =
-                                    qualify_type_name_with_imports(ty_root, prefix, use_imports);
-                                fn_return_types.insert(
-                                    format!("{}::{}", qualified_ty, method.sig.ident),
-                                    ret_ty,
-                                );
-                            }
-                        }
-                    }
-                    // Impl-block associated consts — `impl Foo { const
-                    // CONST_BIT: u32 = 1 << 31; }`. RPython peer:
-                    // `bookkeeper.getdesc(value)` resolves class-level
-                    // descriptors (`bookkeeper.py:329-340 immutablevalue`)
-                    // by host-identity; pyre's static walker registers
-                    // them under `Type::CONST_NAME` so `Type::CONST` /
-                    // `Self::CONST` references can resolve via the
-                    // last-two-segments fallback in
-                    // `lookup_function_return_type` /
-                    // `expr_unary_not_operand_kind`.
-                    if let syn::ImplItem::Const(item_const) = sub {
-                        if let Some(ty) = qualified_full_type_string_with_imports(
-                            &item_const.ty,
-                            prefix,
-                            use_imports,
-                            known_struct_names,
-                            known_trait_names,
-                        ) && let Some(ref ty_root) = self_ty_root
-                        {
-                            let qualified_ty =
-                                qualify_type_name_with_imports(ty_root, prefix, use_imports);
-                            fn_return_types.insert(
-                                format!("{}::{}", qualified_ty, item_const.ident),
-                                ty.clone(),
-                            );
-                            // Bare-key alias for `Self::CONST_BIT`-shape
-                            // references whose qualifier strips to the
-                            // last segment. Mirrors `Item::Const`'s
-                            // file-level registration.
-                            fn_return_types.insert(item_const.ident.to_string(), ty);
-                        }
-                    }
-                }
-            }
-            Item::Trait(trait_def) => {
-                let trait_root = qualify_type_name_with_imports(
-                    &trait_def.ident.to_string(),
-                    prefix,
-                    use_imports,
-                );
-                for sub in &trait_def.items {
-                    if let syn::TraitItem::Fn(method) = sub {
-                        let ret_ty = match &method.sig.output {
-                            syn::ReturnType::Type(_, ty) => {
-                                qualified_full_type_string_with_imports(
-                                    ty,
-                                    prefix,
-                                    use_imports,
-                                    known_struct_names,
-                                    known_trait_names,
-                                )
-                            }
-                            syn::ReturnType::Default => Some("()".to_string()),
-                        };
-                        if let Some(ret_ty) = ret_ty {
-                            fn_return_types
-                                .insert(format!("{}::{}", trait_root, method.sig.ident), ret_ty);
-                        }
-                    }
-                }
-            }
-            Item::Enum(e) => {
-                // RPython sum types are multiple subclasses inheriting from a
-                // common ancestor; each subclass owns its own
-                // `concretetype` field set keyed by the lltype object
-                // identity.  Pyre carries identity as a flat string
-                // table, and the only stable identity for a variant is
-                // the fully-qualified `prefix::Enum::Variant` path.
-                // Earlier drafts also registered the bare `Variant`
-                // fallback and the bare-enum `Enum::Variant` (without
-                // module prefix), but both forms collided across
-                // unrelated enums with the same variant name (e.g.
-                // `a::Foo::Empty` vs `b::Foo::Empty`, or
-                // `Foo::Empty` vs `Bar::Empty`). Register only the
-                // fully-qualified key. `field_type` accepts a shorter
-                // registered key for callers with extra crate prefixes
-                // only when that suffix match is unique. Tuple/unit
-                // variants carry no named fields and need no entry.
-                let bare_enum = e.ident.to_string();
-                let qualified_enum = if prefix.is_empty() {
-                    bare_enum.clone()
-                } else {
-                    format!("{}::{}", prefix, bare_enum)
-                };
-                for variant in &e.variants {
-                    if let syn::Fields::Named(named) = &variant.fields {
-                        let var_name = variant.ident.to_string();
-                        let fields: Vec<(String, String)> = named
-                            .named
-                            .iter()
-                            .filter_map(|f| {
-                                let field_name = f.ident.as_ref()?.to_string();
-                                let field_type = qualified_full_type_string_with_imports(
-                                    &f.ty,
-                                    prefix,
-                                    use_imports,
-                                    known_struct_names,
-                                    known_trait_names,
-                                )?;
-                                Some((field_name, field_type))
-                            })
-                            .collect();
-                        struct_fields
-                            .fields
-                            .insert(format!("{}::{}", qualified_enum, var_name), fields);
-                    }
-                }
-            }
-            Item::Mod(m) => {
-                if let Some((_, ref sub_items)) = m.content {
-                    let mod_prefix = if prefix.is_empty() {
-                        m.ident.to_string()
-                    } else {
-                        format!("{}::{}", prefix, m.ident)
-                    };
-                    collect_fields_and_returns(
-                        sub_items,
-                        &mod_prefix,
-                        use_imports,
-                        known_struct_names,
-                        known_trait_names,
-                        struct_fields,
-                        fn_return_types,
-                        immutable_fields,
-                    );
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-/// Walk the statements of a fn body and register the return types of
-/// any nested `fn` items declared inside.  Used by `collect_fields_and_
-/// returns`'s `Item::Fn` arm so the classifier sees nested-fn
-/// signatures.  Recurses through `Stmt::Item(Item::Fn(_))` only
-/// (nested-mod / nested-impl inside a fn body are vanishingly rare in
-/// pyre source and would require their own qualified prefix).
-fn collect_nested_fn_returns(
-    stmts: &[syn::Stmt],
-    prefix: &str,
-    use_imports: &HashMap<String, String>,
-    known_struct_names: &std::collections::HashSet<String>,
-    known_trait_names: &std::collections::HashSet<String>,
-    fn_return_types: &mut HashMap<String, String>,
-) {
-    for stmt in stmts {
-        if let syn::Stmt::Item(Item::Fn(nested)) = stmt {
-            let ret_ty = match &nested.sig.output {
-                syn::ReturnType::Type(_, ty) => qualified_full_type_string_with_imports(
-                    ty,
-                    prefix,
-                    use_imports,
-                    known_struct_names,
-                    known_trait_names,
-                ),
-                syn::ReturnType::Default => Some("()".to_string()),
-            };
-            if let Some(ret_ty) = ret_ty {
-                let key = if prefix.is_empty() {
-                    nested.sig.ident.to_string()
-                } else {
-                    format!("{}::{}", prefix, nested.sig.ident)
-                };
-                fn_return_types.entry(key).or_insert(ret_ty);
-            }
-            collect_nested_fn_returns(
-                &nested.block.stmts,
-                prefix,
-                use_imports,
-                known_struct_names,
-                known_trait_names,
-                fn_return_types,
-            );
-        }
-    }
-}
+pub use crate::front::syn_metadata::collect_program_metadata as collect_program_metadata_pub;
 
 
 /// Public entry for building a graph from a single function AST node.
